cmd/uplinkng: allow removing pending objects

+++ cmd/uplinkng/ulfs/local.go
@@ -55,7 +55,11 @@ func (l *Local) Create(ctx context.Context, path string) (WriteHandle, error) {
 }
 
 // Remove unlinks the file at the path. It is not an error if the file does not exist.
-func (l *Local) Remove(ctx context.Context, path string) error {
+func (l *Local) Remove(ctx context.Context, path string, opts *RemoveOptions) error {
+	if opts.isPending() {
+		return nil
+	}
+
 	if err := os.Remove(path); os.IsNotExist(err) {
 		return nil
 	} else if err != nil {
@@ -65,9 +69,13 @@ func (l *Local) Remove(ctx context.Context, path string) error {
 	return nil
 }
 
-// ListObjects returns an ObjectIterator listing files and directories that have string prefix
+// List returns an ObjectIterator listing files and directories that have string prefix
 // with the provided path.
-func (l *Local) ListObjects(ctx context.Context, path string, recursive bool) (ObjectIterator, error) {
+func (l *Local) List(ctx context.Context, path string, opts *ListOptions) (ObjectIterator, error) {
+	if opts.isPending() {
+		return emptyObjectIterator{}, nil
+	}
+
 	prefix := path
 	if idx := strings.LastIndex(path, "/"); idx >= 0 {
 		prefix = path[:idx+1]
@@ -80,7 +88,7 @@ func (l *Local) ListObjects(ctx context.Context, path string, recursive bool) (O
 	prefix += string(filepath.Separator)
 
 	var files []os.FileInfo
-	if recursive {
+	if opts.isRecursive() {
 		err = filepath.Walk(prefix, func(path string, info os.FileInfo, err error) error {
 			if err == nil && !info.IsDir() {
 				rel, err := filepath.Rel(prefix, path)
@@ -112,7 +120,7 @@ func (l *Local) ListObjects(ctx context.Context, path string, recursive bool) (O
 	})
 
 	var trim ulloc.Location
-	if !recursive {
+	if !opts.isRecursive() {
 		trim = ulloc.NewLocal(prefix)
 	}
